Move the api request and search results in separate files.
