+++ src/pages/HomePage/HomePage.tsx
@@ -1,33 +1,21 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import SearchBar from '../../components/SearchBar/SearchBar';
+import SearchResults from '../../components/SearchResults/SearchResults';
 import { SearchResult } from '../../data';
+import { searchShows } from '../../utils/requests';
 import styles from './HomePage.module.css';
 
-const API_BASE_URL = 'https://api.tvmaze.com';
-
-export const searchShows = async (query: string) => {
-  const response = await fetch(`${API_BASE_URL}/search/shows?q=${query}`);
-  if (!response.ok) {
-    throw new Error('Failed to fetch shows');
-  }
-  return response.json();
-};
-
-const HomePage: React.FC = () => {
+const HomePage = () => {
   const [results, setResults] = useState<SearchResult[]>([]);
-  const handleSearch = (query: string) => {
-    searchShows(query).then((data) => {
-      setResults(data);
-    });
+
+  const handleSearch = async (query: string) => {
+    const data = await searchShows(query);
+    setResults(data);
   };
   return (
     <div className={styles.homePage}>
       <SearchBar onSearch={handleSearch} />
-      <ul>
-        {results.map((result) => (
-          <li key={result.show.id}>{result.show.name}</li>
-        ))}
-      </ul>
+      <SearchResults results={results} />
     </div>
   );
 };
